+++ dependency-reduced-pom.xml
@@ -54,7 +54,7 @@
       </plugin>
       <plugin>
         <artifactId>maven-compiler-plugin</artifactId>
-        <version>3.14.1</version>
+        <version>3.15.0</version>
         <configuration>
           <compilerArgs>-verbose</compilerArgs>
           <release>21</release>
@@ -78,7 +78,7 @@
       </plugin>
       <plugin>
         <artifactId>maven-shade-plugin</artifactId>
-        <version>3.6.1</version>
+        <version>3.6.2</version>
         <executions>
           <execution>
             <phase>package</phase>
@@ -115,7 +115,7 @@
       </plugin>
       <plugin>
         <artifactId>maven-dependency-plugin</artifactId>
-        <version>3.9.0</version>
+        <version>3.10.0</version>
         <executions>
           <execution>
             <id>copy-dependencies</id>
@@ -131,7 +131,7 @@
       </plugin>
       <plugin>
         <artifactId>maven-resources-plugin</artifactId>
-        <version>3.4.0</version>
+        <version>3.5.0</version>
         <configuration>
           <encoding>UTF-8</encoding>
           <propertiesEncoding>${project.build.sourceEncoding}</propertiesEncoding>
@@ -145,7 +145,7 @@
       </plugin>
       <plugin>
         <artifactId>maven-surefire-plugin</artifactId>
-        <version>3.5.4</version>
+        <version>3.5.5</version>
         <configuration>
           <argLine>-Dfile.encoding=UTF-8</argLine>
           <parallel>false</parallel>
@@ -269,7 +269,7 @@
           <vmArgs>-Dtinylog.writer.level=trace -Xms256m -Xmx1024m</vmArgs>
           <licenseFile>${project.basedir}/LICENSE</licenseFile>
           <url>https://github.com/fransjacobs/model-railway</url>
-          <createZipball>true</createZipball>
+          <createZipball>false</createZipball>
           <createTarball>false</createTarball>
           <assetsDir>${project.basedir}/assets</assetsDir>
           <additionalResources>
@@ -291,7 +291,7 @@
       <plugin>
         <groupId>org.codehaus.mojo</groupId>
         <artifactId>license-maven-plugin</artifactId>
-        <version>2.7.0</version>
+        <version>2.7.1</version>
         <executions>
           <execution>
             <id>aggregate-download-licenses</id>
@@ -392,7 +392,7 @@
           <plugin>
             <groupId>org.codehaus.mojo</groupId>
             <artifactId>license-maven-plugin</artifactId>
-            <version>2.4.0</version>
+            <version>2.7.1</version>
             <executions>
               <execution>
                 <id>aggregate-download-licenses</id>
@@ -410,7 +410,7 @@
           <plugin>
             <groupId>org.codehaus.mojo</groupId>
             <artifactId>license-maven-plugin</artifactId>
-            <version>2.7.0</version>
+            <version>2.7.1</version>
             <executions>
               <execution>
                 <id>aggregate-download-licenses</id>
@@ -436,7 +436,7 @@
     <dependency>
       <groupId>org.junit.jupiter</groupId>
       <artifactId>junit-jupiter-api</artifactId>
-      <version>6.0.1</version>
+      <version>6.0.3</version>
       <scope>test</scope>
       <exclusions>
         <exclusion>
@@ -460,7 +460,7 @@
     <dependency>
       <groupId>org.junit.jupiter</groupId>
       <artifactId>junit-jupiter-params</artifactId>
-      <version>6.0.1</version>
+      <version>6.0.3</version>
       <scope>test</scope>
       <exclusions>
         <exclusion>
@@ -476,7 +476,7 @@
     <dependency>
       <groupId>org.junit.vintage</groupId>
       <artifactId>junit-vintage-engine</artifactId>
-      <version>6.0.1</version>
+      <version>6.0.3</version>
       <scope>test</scope>
       <exclusions>
         <exclusion>
