+++ pom.xml
@@ -85,7 +85,7 @@
         <dependency>
             <groupId>com.twelvemonkeys.imageio</groupId>
             <artifactId>imageio-batik</artifactId>
-            <version>3.12.0</version>
+            <version>3.13.1</version>
             <exclusions>
                 <exclusion>
                     <groupId>org.apache.commons</groupId>
@@ -113,13 +113,13 @@
         <dependency>
             <groupId>org.junit.jupiter</groupId>
             <artifactId>junit-jupiter-api</artifactId>
-            <version>6.0.1</version>
+            <version>6.0.3</version>
             <scope>test</scope>
         </dependency>
         <dependency>
             <groupId>org.junit.jupiter</groupId>
             <artifactId>junit-jupiter-params</artifactId>
-            <version>6.0.1</version>
+            <version>6.0.3</version>
             <scope>test</scope>
             <exclusions>
                 <exclusion>
@@ -131,7 +131,7 @@
         <dependency>
             <groupId>org.junit.vintage</groupId>
             <artifactId>junit-vintage-engine</artifactId>
-            <version>6.0.1</version>
+            <version>6.0.3</version>
             <scope>test</scope>
         </dependency>
         <dependency>
@@ -167,12 +167,12 @@
         <dependency>
             <groupId>com.formdev</groupId>
             <artifactId>flatlaf</artifactId>
-            <version>3.6.2</version>
+            <version>3.7.1</version>
         </dependency>
         <dependency>
             <groupId>org.httprpc</groupId>
             <artifactId>sierra</artifactId>
-            <version>3.5.2</version>
+            <version>4.3.1</version>
         </dependency>     
         <dependency>
             <groupId>eu.hansolo</groupId>
@@ -202,12 +202,12 @@
         <dependency>
             <groupId>ch.qos.logback</groupId>
             <artifactId>logback-classic</artifactId>
-            <version>1.5.25</version>
+            <version>1.5.32</version>
         </dependency>
         <dependency>
             <groupId>org.netbeans.external</groupId>
             <artifactId>AbsoluteLayout</artifactId>
-            <version>RELEASE280</version>
+            <version>RELEASE290</version>
         </dependency>
     </dependencies>
     <build>        
@@ -222,7 +222,7 @@
             </plugin>
             <plugin>
                 <artifactId>maven-compiler-plugin</artifactId>
-                <version>3.14.1</version>
+                <version>3.15.0</version>
                 <configuration>
                     <!--compilerArgs>-Xlint:unchecked,-verbose</compilerArgs-->
                     <compilerArgs>-verbose</compilerArgs>
@@ -249,7 +249,7 @@
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-shade-plugin</artifactId>
-                <version>3.6.1</version>
+                <version>3.6.2</version>
                 <executions>
                     <execution>
                         <phase>package</phase>
@@ -287,7 +287,7 @@
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-dependency-plugin</artifactId>
-                <version>3.9.0</version>
+                <version>3.10.0</version>
                 <executions>
                     <execution>
                         <id>copy-dependencies</id>
@@ -305,7 +305,7 @@
             </plugin>
             <plugin>
                 <artifactId>maven-resources-plugin</artifactId>
-                <version>3.4.0</version>              
+                <version>3.5.0</version>              
                 <configuration>
                     <encoding>UTF-8</encoding>
                     <propertiesEncoding>${project.build.sourceEncoding}</propertiesEncoding>
@@ -320,7 +320,7 @@
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-surefire-plugin</artifactId>
-                <version>3.5.4</version>
+                <version>3.5.5</version>
                 <configuration>
                     <!--testFailureIgnore>false</testFailureIgnore-->
                     <argLine>-Dfile.encoding=UTF-8</argLine>
@@ -391,7 +391,7 @@
                     <vmArgs>-Dtinylog.writer.level=trace -Xms256m -Xmx1024m</vmArgs> 
                     <licenseFile>${project.basedir}/LICENSE</licenseFile>
                     <url>https://github.com/fransjacobs/model-railway</url>
-                    <createZipball>true</createZipball>
+                    <createZipball>false</createZipball>
                     <createTarball>false</createTarball>
                     <assetsDir>${project.basedir}/assets</assetsDir>
                     <additionalResources>
@@ -540,7 +540,7 @@
             <plugin>
                 <groupId>org.codehaus.mojo</groupId>
                 <artifactId>license-maven-plugin</artifactId>
-                <version>2.7.0</version>
+                <version>2.7.1</version>
                 <configuration>
                     <licensesOutputDirectory>${project.build.directory}/generated-resources/licenses</licensesOutputDirectory>
                     <useRepositoryContent>true</useRepositoryContent>
@@ -670,7 +670,7 @@
                     <plugin>
                         <groupId>org.codehaus.mojo</groupId>
                         <artifactId>license-maven-plugin</artifactId>
-                        <version>2.4.0</version>
+                        <version>2.7.1</version>
                         <executions>
                             <execution>
                                 <id>aggregate-download-licenses</id>
@@ -688,7 +688,7 @@
                     <plugin>
                         <groupId>org.codehaus.mojo</groupId>
                         <artifactId>license-maven-plugin</artifactId>
-                        <version>2.7.0</version>
+                        <version>2.7.1</version>
                         <executions>
                             <execution>
                                 <id>aggregate-download-licenses</id>
