+++ src/main/java/jcs/ui/panel/DispatcherTablePanel.form
@@ -41,6 +41,7 @@
       <SubComponents>
         <Component class="javax.swing.JTable" name="dispatcherTable">
           <Properties>
+            <Property name="autoCreateRowSorter" type="boolean" value="true"/>
             <Property name="model" type="javax.swing.table.TableModel" editor="org.netbeans.modules.form.RADConnectionPropertyEditor">
               <Connection component="locomotiveDispatcherTableModel" type="bean"/>
             </Property>
